让张力系统支持小数位，汉化张力调整面板，显示小数后3位

+++ Scripts/Customs/Mikes Scripts/Utilities/Tension System/AugmentedParagonInfo.cs
@@ -65,7 +65,7 @@ namespace Server.Mobiles
         /// </summary>
         public double GetBuffMultiplier(double baseValue)
         {
-            int tension = TensionManager.Tension;
+            double tension = TensionManager.Tension;
             double scaleFactor = 1.0 + (tension / 50000.0);
             return baseValue * scaleFactor;
         }
